Update code to delete image when product is deleted or changed

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator');
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res) => {
   res.render('admin/edit-product', {
@@ -113,6 +114,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save()
@@ -146,7 +148,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.session.user })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        throw new Error('Product not Found!');
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.session.user });
+    })
     .then(() => {
       res.redirect('/admin/products');
     })
